% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPathwayEnrichment.R
\name{FindPathwayEnrichment}
\alias{FindPathwayEnrichment}
\title{Find Reactome pathways enriched from proximal binding}
\usage{
FindPathwayEnrichment(distanceFrame, minPathwayGenes = 5, numCores = 4)
}
\arguments{
\item{distanceFrame}{A dataframe resulting from the output of the GeneToPeakDist function. Should have two columns with the first having ENSEMBL gene IDs and the second the distance to the nearest ChIP peak.}

\item{minPathwayGenes}{A numeric indicating the minimum number of genes in a pathway required for inclusion in pathway analysis. Defaults to 5.}

\item{numCores}{A numeric indicating the number of cores the function can use via R's parallel package. Defaults to 4.}
}
\value{
Returns a vector of one-sided Wilcoxon p-values with the alternate hypothesis of pathway genes enriched for shorter distances to the nearest ChIP-peak relative to all non-pathway genes.
}
\description{
Find Reactome pathways enriched from proximal binding
}
\note{
Reactome

Terms of Use: http://www.reactome.org/pages/about/license-agreement/
}

